Exercise controller updated

+++ ITTWEB-Assignment6-FitnessApp/Models/User.cs
@@ -4,7 +4,7 @@ namespace ITTWEB_Assignment6_FitnessApp.Models
 {
     public class User
     {
-        public double Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
         public List<Workout> Workouts { get; set; }
